start wallet command refactoring
