+++ wallet/src/lib.rs
@@ -55,6 +55,7 @@ pub mod libwallet;
 pub mod lmdb_wallet;
 mod node_clients;
 mod types;
+pub mod command;
 
 pub use adapters::{
 	FileWalletCommAdapter, HTTPWalletCommAdapter, NullWalletCommAdapter, WalletCommAdapter,
